backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai-whisper==20231117
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
noisereduce==3.0.0
soundfile==0.12.1
